cleanup useDragItem hook logic
